Split behavior ref role tests
